Use load tests protocol to run the test suite
Since we no longer require Python 2, we can make use of the load tests
protocol and automatic discovery to find all tests we want to run,
rather than constructing a list of TestSuites by hand. As a consequence,
I've had to fiddle with two test classes so they are not discovered and
run. Drive-by the tox environment list as well.