feat(test): Add traceability requirements
The intention of CCT-1063 is to add a gating.el10 job for python-iniparse.
Before that it would be good to have the traceability matrix to
be able to report our tests to Polarion. This PR therefore adds the
neccessary docstrings for Betelgeuse to each test. It also adds two
github workflows - checks for betelgeuse and for testimony so that
on every PR we have a check that the tests have these requirements.